(refactor) created stylesheet in lib, refactored SignUp to import styles
