+++ mobile/components/Signup.js
@@ -5,6 +5,7 @@ var windowSize = Dimensions.get('window');
 var Login = require('./Login');
 var Main = require('./Main');
 var utils = require('../lib/utility');
+var styles = require('../lib/stylesheet');
 
 var {
   AppRegistry,
@@ -135,80 +136,80 @@ class Signup extends Component {
 
 };
 
-var styles = StyleSheet.create({
-  container: {
-    flexDirection: 'column',
-    flex: 1,
-    backgroundColor: 'transparent'
-  },
-  bg: {
-      position: 'absolute',
-      left: 0,
-      top: 0,
-      width: windowSize.width,
-      height: windowSize.height
-  },
-  header: {
-      justifyContent: 'center',
-      alignItems: 'center',
-      flex: .5,
-      backgroundColor: 'transparent'
-  },
-  mark: {
-      width: 150,
-      height: 150
-  },
-  signup: {
-      backgroundColor: '#FF3366',
-      padding: 20,
-      alignItems: 'center',
-      marginTop: -8.5
-  },
-  login: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    flex: .15
-  },
-  inputs: {
-      marginTop: 200,
-      marginBottom: 10,
-      flex: .25
-  },
-  inputPassword: {
-      marginLeft: 10,
-      width: 20,
-      height: 21
-  },
-  inputUsername: {
-    marginLeft: 10,
-    width: 20,
-    height: 20
-  },
-  inputContainer: {
-      padding: 10,
-      borderWidth: 1,
-      borderBottomColor: '#CCC',
-      borderColor: 'transparent'
-  },
-  input: {
-      position: 'absolute',
-      left: 61,
-      top: 12,
-      right: 0,
-      height: 20,
-      fontSize: 14
-  },
-  forgotContainer: {
-    alignItems: 'flex-end',
-    padding: 15,
-  },
-  greyFont: {
-    color: '#D8D8D8'
-  },
-  whiteFont: {
-    color: '#FFF'
-  }
-})
+// var styles = StyleSheet.create({
+//   container: {
+//     flexDirection: 'column',
+//     flex: 1,
+//     backgroundColor: 'transparent'
+//   },
+//   bg: {
+//       position: 'absolute',
+//       left: 0,
+//       top: 0,
+//       width: windowSize.width,
+//       height: windowSize.height
+//   },
+//   header: {
+//       justifyContent: 'center',
+//       alignItems: 'center',
+//       flex: .5,
+//       backgroundColor: 'transparent'
+//   },
+//   mark: {
+//       width: 150,
+//       height: 150
+//   },
+//   signup: {
+//       backgroundColor: '#FF3366',
+//       padding: 20,
+//       alignItems: 'center',
+//       marginTop: -8.5
+//   },
+//   login: {
+//     justifyContent: 'center',
+//     alignItems: 'center',
+//     flex: .15
+//   },
+//   inputs: {
+//       marginTop: 200,
+//       marginBottom: 10,
+//       flex: .25
+//   },
+//   inputPassword: {
+//       marginLeft: 10,
+//       width: 20,
+//       height: 21
+//   },
+//   inputUsername: {
+//     marginLeft: 10,
+//     width: 20,
+//     height: 20
+//   },
+//   inputContainer: {
+//       padding: 10,
+//       borderWidth: 1,
+//       borderBottomColor: '#CCC',
+//       borderColor: 'transparent'
+//   },
+//   input: {
+//       position: 'absolute',
+//       left: 61,
+//       top: 12,
+//       right: 0,
+//       height: 20,
+//       fontSize: 14
+//   },
+//   forgotContainer: {
+//     alignItems: 'flex-end',
+//     padding: 15,
+//   },
+//   greyFont: {
+//     color: '#D8D8D8'
+//   },
+//   whiteFont: {
+//     color: '#FFF'
+//   }
+// })
 module.exports = Signup;
